fix: address PR review comments from Copilot
- Fix UTF-8 panic in print_summary() by using truncate_preview()
- Harden parse_marker() to only match inside {code:title=hearken-metadata}
  blocks, preventing false positives from log content
- Reconstruct metadata block delimiters in ADF text extraction
- Use i64::MAX as usize instead of usize::MAX for pattern limit
- Add explicit tokio "time" feature for rate-limit sleep
- Fix "bidirectional" wording to "one-way" in ARCHITECTURE.md
- Remove non-existent user TOML key from README config example
- Fix changelog version to 0.3.0 matching crate versions
- Remove undocumented basic auth fallback from design spec

+++ hearken-jira/src/filter.rs
@@ -24,7 +24,7 @@ pub fn filter_patterns(
     synced_pattern_ids: &HashSet<i64>,
     anomaly_ids: Option<&HashSet<i64>>,
 ) -> Result<Vec<FilteredPattern>, hearken_storage::StorageError> {
-    let all_patterns = storage.get_all_patterns_ranked(usize::MAX, None, None)?;
+    let all_patterns = storage.get_all_patterns_ranked(i64::MAX as usize, None, None)?;
 
     let need_tags = options.tags.is_some() || options.exclude_tags.is_some();
     let tag_map: HashMap<i64, Vec<String>> = if need_tags {
